pyproject.toml
[project]
name = "intfeat"
version = "0.1.0"
description = "Research on integer features"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "jupyter>=1.0.0",
 "numpy>=2.3.0",
 "scipy>=1.16.0",
 "pandas>=2.3.0",
 "pyyaml>=6.0.2",
 "matplotlib>=3.10.0",
 "tqdm>=4.67.1",
 "ucimlrepo>=0.0.7",
 "rpy2>=3.6.3",
 "seaborn>=0.11.0",
 "optuna>=4.5.0",
 "polars>=1.33.0",
 "tensordict>=0.10.0",
 "torch>=2.8.0",
 "torchcurves>=0.2.0",
 "statsmodels>=0.14.5",
 "scikit-learn>=1.7.0",
 "cvxpy>=1.7.0",
 "great-tables>=0.18.0",
 "pytest>=8.3.0",
]
